+++ kernel/kernel-masterdata-service/src/main/java/io/mosip/kernel/masterdata/service/LocationService.java
@@ -78,7 +78,7 @@ public interface LocationService {
 	 *            - location code
 	 * @param langCode
 	 *            - language code
-	 * @return location response dto
+	 * @return {@link LocationResponseDto}
 	 */
 	public LocationResponseDto getImmediateChildrenByLocCodeAndLangCode(String locCode, String langCode);
 
@@ -91,12 +91,14 @@ public interface LocationService {
 	 * @return map contain key as parentCode and value as List of Location
 	 * 
 	 */
-	public Map<Integer, List<Location>> getLocationByLangCodeAndHierarchyLevel(String langCode, Integer hierarchyLevel);
-	
+	public Map<Short, List<Location>> getLocationByLangCodeAndHierarchyLevel(String langCode, Short hierarchyLevel);
+
 	/**
 	 * checks whether the given location name is valid or not
+	 * 
 	 * @param locationName
-	 * @return StatusResponseDto
+	 *            -location name
+	 * @return {@link StatusResponseDto}
 	 */
 	public StatusResponseDto validateLocationName(String locationName);
